Выполняет домашнее задание Модуляция
